pyproject.toml
[tool.poetry]
name = "meteocatpy"
version = "0.0.6"
description = "Script para obtener datos meteorológicos de la API de Meteocat"
authors = ["figorr <jdcuartero@yahoo.es>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://gitlab.com/figorr/meteocatpy"
keywords = ['meteocatpy', 'meteocat']


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
simplejson = "^3.19.3"
setuptools = "^75.5.0"
python-semantic-release = "^7.28.1"
twine = ">=3,<4"
aiofiles = "^24.1.0"
voluptuous = "^0.15.2"
geopy = "^2.4.1"
tzdata = "^2024.2"
aiohttp = ">=3.10.11,<4.0.0"
diskcache = "^5.6.3"


[tool.semantic_release]
branch = "master"  # Define la rama desde la que se harán los lanzamientos
version_source = "commit"  # Determina la versión basada en los commits
version_variable = "meteocatpy/version.py:__version__"  # Ubicación de la variable de versión
changelog = { file = "CHANGELOG.md" }  # Genera el changelog en el archivo especificado
upload_to_pypi = false  # Sube automáticamente la versión a PyPI
ci = false  # Cambiar a false si no usas CI/CD


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dev-dependencies]
aresponses = "*"
detox = "*"
flake8 = "*"
mypy = "*"
pydocstyle = "*"
pylint = "*"
pytest-aiohttp = "*"
pytest-cov = "*"
python-semantic-release = "^7.28.1"
tox = "*"
safety = ">=1.8.7"
black = { version = ">19.10b0", allow-prereleases = true }
Sphinx = ">=3.5.0,<7.0.0"
autoapi = ">=2.0.1"
sphinx-rtd-theme = "^0.5.1"
m2r2 = ">=0.2.7"
tomlkit = ">=0.7.0"
sphinx-autoapi = ">=1.7.0"
sphinx-copybutton = ">=0.3.1"
pipdeptree = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pyupgrade = "^3.4.0"
pre-commit = "^3.3.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
syrupy = "^4.7.2"
aioresponses = "^0.7.7"
